playerdata

+++ TheAlbianTimes/Assets/Scripts/GameState/GameState.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GameStatePlayerData
 {
-    public int money = 1000;
+    public float money = 1000f;
     public int staff = 10;
     public int reputation = 50;
 }
